handle accidental macro start and repeat macro binding
This fixes the problem of lot of accidental macro recording start with
`q` since many terminal tools have q as quit(tig, nnn). The fix is to
remap it to Q and set nop to q. We can still trigger ex mode with gQ.

'\' is mapped to @@ to repeat last macro.