Bind CocSearch
CocSearch runs rg and provides all the matching results in one buffer. A
macro can be run on the buffer or simply we can run %s command. And then on
saving the buffer, it saves the changes in respective files. In the same
buffer we can go on the file name and clicking enter opens the file in a
new buffer.

We can also pass optional rg arguments like -A20 to show the context
around the match.