add commit template to global git config
A commit template makes the person committing more aware of the
information that needs to be provided in the commit message. This allows
to be more explicit in explaining the problem/commit etc.

This commit also includes formatting of the gitconfig file to the
default format of tabs as indentation. Also add back git commit binding
to c in tigrc