proper dotfiles config using rsync
manually copying dotfiles was not elegant. Maintaining symlinks was a
bit complex sometimes. rsync seems to be a much better solution. Added a
cron job that runs rsync on dotfiles every 10min. rsync is not
expensive. It will sync only those files that has changed. Hence we dont
need all those enteries in gitignore. This solution is better because
not fish does not treat every path under home directory as part of the
git project.